requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
SQLAlchemy-Utils==0.41.2
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
pytest==7.4.2
httpx==0.24.1
pytest-asyncio==0.22.0
pytest-cov==4.1.0
coverage==7.4.2

# Dependências adicionais podem ser listadas aqui